Added toggleable buttons, updated pushed button and open gate graphics, added TODO file
Buttons can now be set as toggleable, which means that they will become unpressed when something is no longer on top of them.

Updated the pushed button graphics.  Buttons will now be smaller when they are in the pushed state.

Updated the open gate graphics.  Gates will now be translucent and unable to collide while their button is pushed.

Added a TODO.txt file to list known bugs.